requirements.txt
rplidar-roboticia
plotly
pyvista